requirements.txt
glad==0.1.34
glfw==2.1.0
numpy==1.20.1
PyOpenGL==3.1.5
PyOpenGL-accelerate==3.1.5
